block create acc admin from user

+++ api/src/main/java/com/myshop/api/service/user/UserServiceImpl.java
@@ -131,6 +131,9 @@ public class UserServiceImpl extends CRUDBaseServiceImpl<UserInfo, UserRequest, 
         if (role == null) {
             return UserResponse.builder().status(false).message("Role is not exists, please check again").build();
         }
+        if (role.getName() == "ROLE_ADMIN") {
+            return UserResponse.builder().status(false).message("You do not have permission to create an admin account").build();
+        }
         account = Account.builder().username(userRequest.getUserName()).deleteFlag(false).role(role).email(userRequest.getEmail())
                 .password(passwordEncoder.encode(userRequest.getPassword() + Constants.SALT_DEFAULT)).build();
         UserInfo userInfo = UserInfo.builder().firstName(userRequest.getFirstName())
